refactor(orchestration): extract message orchestrator and router state
Split the 795-line src/index.ts god object into three focused modules:

- src/router-state.ts — RouterState class owns lastTimestamp, sessions,
  registeredGroups, lastAgentTimestamp with load/save; getAvailableGroups
  is now a pure function that takes registeredGroups as a param
- src/message-orchestrator.ts — createMessageOrchestrator() factory owns
  processGroupMessages, runAgent, startMessageLoop, recoverPendingMessages
- src/index.ts — thin initialization (298 lines, down from 795)

routing.test.ts updated: getAvailableGroups is pure so _setRegisteredGroups
is no longer needed — tests pass registeredGroups directly.

All 552 tests pass.
